Advanced CT Cardiac Services at Butler Medical Imaging. (08) 9544 3999
What is CT Cardiac?

A cardiac CT, or CT Coronary Angiogram, is a specialised non-invasive procedure which uses a mixture of x-ray technology, computer systems and a contrast medium to produce detailed three dimensional images of the heart, arteries and blood vessels.

What part of the anatomy is examined by a cardiac CT?

Cardiac CTs are used specifically to examine the heart.

What conditions can be diagnosed by a cardiac CT?

Cardiac CTs are typically used to detect and aid in the diagnosis of coronary heart disease.

Why is a cardiac CT required instead of a Coronary Angiogram?

A cardiac CT is a non-invasive procedure, which makes it ideal for checking whether a patient with low risk heart-related issues or symptoms has a coronary artery disease. Doctors also use cardiac CTs to check for blockages and damage to the coronary arteries. Doctors may also require patients to get a cardiac CT scan before or after other coronary procedures.

What consequences are there if the suspected condition is undiagnosed or untreated?

Coronary artery disease is a serious illness, which if left untreated can lead to the development of other serious conditions including:

·         Heart attack

·         Arrhythmia

·         Heart failure

·         Death

·         Preparation for CT Cardiac

What should you tell us when making a booking for a cardiac CT?

·         If you are pregnant

·         If you are diabetic

·         If you have kidney problems

·         If you have multiple severe allergies, or a previous allergy to IV contrast

·         You will usually be asked to have a blood test to check your kidney function before the test, and definitely if you are diabetic.

Warmth or heat What Special Diet is required Before a cardiac CT?

You may be instructed to fast after breakfast and avoid any caffeine for up to 24 hours before your procedure.

What to bring for a cardiac CT?

Your doctor may ask you to bring any previous scans or x-rays to the procedure, along with any medications you’re currently taking.

What to wear for a cardiac CT?

You will be asked to remove any metallic items, such as jewellery, watches, eyewear and belts, as these items can show up on the CT scans. You will be asked to remove any clothing obstructing the area to be scanned and given a medical gown to wear instead, so you should try and wear something that is comfortable and easy to remove.

How long will it take?

The pre-requisites of the CT Cardiac take a long time in processes such as stabilizing heart rate and doing the initial heart scans otherwise the actual process doesn’t take more than 10 to 15 minutes.

What does a cardiac CT involve?

After arrival at the Radiology Department, you will be placed on a trolley to have your heart rate (HR) measured. If it is too high to perform the test, you will be given medications to lower your HR, and it will be rechecked after 1/2 – 1 hour. Once your HR is at a good level, a nurse, doctor or radiographer will put a cannula in your arm, usually in the inside of the elbow. You will be taken to the CT Scan room and asked to lay on the CT bed. The radiographers will connect you to the CT Scan ECG and explain the next part of the test to you. You will be asked to hold your breath during all of the scans – this is very important. A couple of preliminary scans will be done and then the radiographer will start the injection of contrast, and the CT Coronary scan will start shortly after that.

You may experience the following sensations as the contrast agent moves through the body:

·         Mild burning

·         A metallic taste or smell

·         A feeling as though you have passed urine

·         Post cardiac CT instructions

What are the Recovery Details?

You may be asked to stay in the hospital for up to an hour after the cardiac CT, to ensure that there are no side effects or reactions to the contrast medium or the medications given to lower your HR.

After the procedure you will be able to return to normal activity the next day. The results of the scan will be sent to your doctor, who will discuss them with you during your next appointment.

How much does a CT Cardiac cost?

At Butler Medical Imaging, we pride ourselves on our ability to deliver first class services within the health system using Bulk billing methodology. This means that your consultation and procedure will be bulk billed if you have your Medicare card and a practitioner’s referral with you, and you will have no out of pocket expenses.

Do I need an appointment?

Yes, please give us a call at +61 8 9544 3999 and one of our friendly staff members will help you. You can also make a booking by yourself by clicking here.

Can I use my referral anywhere?

If you have a referral to another imaging practice or organization, we can assist you as we can accept all referrals and if you are still in doubt, please seek advice from your referring medical professional.

Why Choose Butler Medical Imaging?

We are local and a family-owned medical clinics that offers the highest quality medical and Imaging Services in the Butler Medical Imaging region. We are the leaders in Bulk Billing with the real benefit of no ‘Out of Pocket’ expenses for you.

Your Next Step

Please ensure that you have your Medicare card and referral with you and pop into our location for your consultation. If you have any questions at all, please feel free to contact us at +61 8 9544 3999 as we are here to help.

Contact Us:

Phone: (08) 9544 3999

Email: info@butlerimaging.com.au

Address: Shop 29/150, Camborne Pkwy, Butler, WA, 6036

Website: www.butlerimaging.com.au